+++ src/Enums/SortDirection.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace HiFolks\DataType\Enums;
+
+enum SortDirection: string
+{
+    case ASC = "asc";
+    case DESC = "desc";
+}